==> 6-paper-tables-figures.Rmd <==
---
title: "Item Formats Paper Tables and Figures"
author: "Benjamin Shear"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Packages

```{r}
library(tidyverse)
library(sjPlot)
library(kableExtra)
```

# Load data

```{r}
# load full dataset for Table 2

full_data <- readRDS(file = "2d-cog-scored-long-allyears-standard-subset.Rds")

# load US DIF estimates and regression models

dif_ests <- readRDS(file = "4b-dif-estimates-data-US-ypart.RDS")
dif_reg <- readRDS(file = "4b-lm-models-US-ypart.rds")

# load DFF model estimates

dff_ests <- read.csv(file = "5b-dff-estimates-contrastcode.csv")
```

# Preliminary statistics

## students per country-year

```{r}
# summarize number of students per year
full_data %>%
  group_by(CNT, YEAR, STUID) %>%
  summarise() %>%
  group_by(CNT, YEAR) %>%
  summarise(N=n()) %>%
  summary()
```

## number of item-year observations per country and unique items

```{r}
# summarize number of item-year and "unique" items per subject
full_data %>%
  group_by(SUB, YEAR, item_id, item_id_short, is_mc) %>%
  summarise() %>%
  group_by(SUB) %>%
  summarise(
    itemyear=length(unique(paste0(item_id,YEAR))),
    item=length(unique(item_id_short)),
    N=n(),
    MC=mean(is_mc)
  ) %>%
  kable() %>% kable_styling()
```

## number of item-year observations

```{r}
# summarize number of item-year-sub observations per country
full_data %>%
  group_by(CNT, YEAR, SUB, item_id) %>%
  summarise() %>%
  group_by(CNT) %>%
  summarise(N=n()) %>%
  summary()
```

## % of items polytomous

```{r}
# percent polytomously scored
full_data %>%
  group_by(YEAR, SUB, item_id, is_poly) %>%
  summarise() %>%
  group_by(SUB) %>%
  summarise(poly=mean(is_poly)) %>%
  kable() %>% kable_styling()
```

## % of all items MC

```{r}
# % of all items MC
full_data %>%
  group_by(YEAR, SUB, item_id, is_mc) %>%
  summarise() %>%
  group_by(SUB) %>%
  summarise(
    N=n(),
    MC=mean(is_mc)
  ) %>%
  kable() %>% kable_styling()

full_data %>%
  group_by(YEAR, item_id, is_mc) %>%
  summarise() %>%
  with(., mean(is_mc))
```

## % of all responses skipped

```{r}
# % responses skipped
mean(full_data$resp_type=="skip")
```

## % of all responses not reached

```{r}
# % responses not reached
mean(full_data$resp_type=="nr")
```

## total number of item responses

```{r}
# total number of item responses
nrow(full_data)
```

## total number of students

```{r}
# total number of students
full_data %>%
  group_by(CNT, YEAR, STUID) %>%
  summarise() %>%
  nrow()
```

## total item-country-year observations

```{r}
# total item-country-year observations
full_data %>%
  group_by(CNT, YEAR, SUB, item_id) %>%
  summarise() %>%
  nrow()
```

## number significant DIF estimates

```{r}
# number of significant DIF estimates?
dif_ests %>%
  group_by(subject, is_mc) %>%
  summarise(
    N=n(),
    Sig=sum(abs(dif/dif_se)>1.96),
    Perc=Sig/N
  ) %>%
  kable() %>% kable_styling()

dif_ests %>%
  group_by(subject) %>%
  summarise(
    N=n(),
    Sig=sum(abs(dif/dif_se)>1.96),
    Perc=Sig/N
  ) %>%
  kable() %>% kable_styling()

dif_ests %>%
  mutate(sig = ifelse(abs(dif/dif_se)>1.96, 1, 0)) %>%
  group_by(subject) %>%
  summarise(
    N=n(),
    Sig=sum(sig),
    Perc=Sig/N,
    SigPos=sum(dif>0 & sig==1),
    SigNeg=sum(dif<0 & sig==1)
  ) %>%
  kable() %>% kable_styling()

```

## omission rate analysis

```{r}
full_data %>%
  group_by(is_mc, GENDER) %>%
  summarise(
    Omit=mean(resp_type!="answer")*100
  ) %>%
  pivot_wider(names_from = "GENDER", values_from="Omit") %>%
  mutate(Diff=(FEMALE-MALE)) %>%
  kable() %>% kable_styling()
```

```{r}
full_data %>%
  group_by(SUB, is_mc, GENDER) %>%
  summarise(
    Omit=mean(resp_type!="answer")*100
  ) %>%
  pivot_wider(names_from = "GENDER", values_from="Omit") %>%
  mutate(Diff=(FEMALE-MALE)) %>%
  kable() %>% kable_styling()
```

```{r}
full_data %>%
  group_by(SUB, YEAR, is_mc, GENDER) %>%
  summarise(
    Omit=mean(resp_type!="answer")*100
  ) %>%
  pivot_wider(names_from = "GENDER", values_from="Omit") %>%
  mutate(Diff=(FEMALE-MALE)) %>%
  kable() %>% kable_styling()
```

### For US Samples:

```{r}
full_data %>%
  filter(CNT=="USA") %>%
  group_by(SUB, is_mc, GENDER) %>%
  summarise(
    Omit=mean(resp_type!="answer")*100
  ) %>%
  pivot_wider(names_from = "GENDER", values_from="Omit") %>%
  mutate(Diff=(FEMALE-MALE)) %>%
  kable() %>% kable_styling()
```

```{r}
full_data %>%
  filter(CNT=="USA") %>%
  group_by(SUB, YEAR, is_mc, GENDER) %>%
  summarise(
    Omit=mean(resp_type!="answer")*100
  ) %>%
  pivot_wider(names_from = "GENDER", values_from="Omit") %>%
  mutate(Diff=(FEMALE-MALE)) %>%
  kable() %>% kable_styling()
```

# Tables

## Table 1

Separate document.

## Table 2

```{r}
dat_usa <- filter(full_data, CNT=="USA") %>%
  mutate(FEMALE=ifelse(GENDER=="FEMALE",1,0))

item_cols <- dat_usa %>%
  group_by(SUB,YEAR,item_id,is_mc) %>%
  summarise(
    x=mean(ypart)
  ) %>%
  group_by(SUB,YEAR) %>%
  summarise(
    NAll=n(),
    NMC=sum(is_mc),
    NCR=sum(is_mc==0),
    PAll=mean(x),
    PMC=mean(x[is_mc==1]),
    PCR=mean(x[is_mc==0])
  ) %>% ungroup()

student_cols <- dat_usa %>%
  group_by(SUB,YEAR,CNT,STUID,FEMALE,AGE,GRADE) %>%
  summarise() %>%
  group_by(SUB,YEAR) %>%
  summarise(
    N=n(),
    `%Female`=mean(FEMALE),
    MnAge=mean(AGE),
    MnGrd=mean(GRADE)
  )

# Next, summarize differences in percent correct scores,
# limiting to students who responded to at least 1 MC and 1 CR

perc_cor_diffs <- dat_usa %>%
  group_by(CNT,SUB,YEAR,STUID,FEMALE) %>%
  summarise(
    N=n(),
    NMC=sum(is_mc),
    NCR=sum(is_mc==0),
    All=mean(ypart),
    MC=mean(ypart[is_mc==1]),
    CR=mean(ypart[is_mc==0])
  ) %>%
  filter(N>=5,NMC>=1,NCR>=1) %>%
  pivot_longer(c("All","MC","CR")) %>%
  group_by(SUB,YEAR,name) %>%
  summarise(
    d=(mean(value[FEMALE==1])-mean(value[FEMALE==0]))/
      ((sd(value[FEMALE==1])+sd(value[FEMALE==0]))/2),
    N=n(),
  ) %>%
  pivot_wider(names_from="name",values_from=c("d","N"))

# Now combine columns together. 

left_join(item_cols, student_cols) %>%
  select(SUB,YEAR,NAll,NCR,NMC,PAll,PCR,PMC,N,`%Female`) %>%
  pivot_longer(cols = -c("SUB","YEAR")) %>%
  unite(subyear, SUB,YEAR) %>%
  pivot_wider(names_from = "subyear", values_from="value") %>%
  bind_rows(., perc_cor_diffs %>%
              select("SUB","YEAR", starts_with("d_")) %>%
              pivot_longer(cols = -c("SUB","YEAR")) %>%
              unite(subyear, SUB,YEAR) %>%
              pivot_wider(names_from = "subyear", values_from="value")) %>%
  arrange(match(name,c(
    "NAll","NCR","NMC",
    "PAll","PCR","PMC",
    "d_All","d_CR","d_MC",
    "N","%Female"))) %>%
  kable(digits=3) %>% kable_styling()
```

## Table 3

```{r}
usedigits <- 3

# tam section #####
tam_section <- dif_ests %>%
  group_by(subject, year, is_mc) %>%
  summarise(
    mn=mean(dif),
    min=min(dif),
    max=max(dif),
    sigma_theta=mean(theta_sd),
    gamma2=(2*mean(main_eff))/mean(theta_sd)
  ) %>%
  group_by(subject, year) %>%
  ungroup() %>%
  pivot_longer(cols = c("mn","min","max","sigma_theta","gamma2")) %>%
  pivot_wider(names_from = c("subject","year"), values_from="value") %>%
  filter(
    !(is_mc==0 & name=="sigma_theta"),
    !(is_mc==0 & name=="gamma2")) %>%
  unite(name, name, is_mc)

tam_section %>%
  kable(digits=usedigits) %>% kable_styling()
```

```{r}
usedigits <- 3

# DFF LLTM section #####
lltm_section <- dff_ests %>%
  filter(CNT=="USA") %>%
  group_by(CNT, SUB, YEAR) %>%
  mutate(
    B1=estimate[term=="Female"]
  ) %>%
  ungroup() %>%
  filter(term=="Female X isCR") %>%
  mutate(
    SUB=ifelse(SUB=="Math","M","R"),
    B3 = format(round(B3,usedigits),nsmall=usedigits),
    B3 = ifelse(B3_p<0.001,
                paste0(B3,"***"),
                ifelse(B3_p<0.01,
                       paste0(B3,"**"),
                       ifelse(B3_p<0.05,
                              paste0(B3,"*"),
                              B3))),
    B3_se=format(round(B3_se,usedigits),nsmall=usedigits),
    d_format=format(round(d_format,usedigits), nsmall=usedigits),
    sdtheta=format(round(sdtheta,usedigits),nsmall=usedigits),
    sditem=format(round(sditem,usedigits),nsmall=usedigits)
  ) %>%
  select(SUB, YEAR, B3,B3_se, d_format, starts_with("sd")) %>%
  pivot_longer(cols = -c("SUB", "YEAR")) %>%
  pivot_wider(names_from = c("SUB","YEAR"), values_from="value")

lltm_section %>%
  kable() %>% kable_styling()
```

## Table 4

```{r}
tab_model(dif_reg,
          dv.labels = c("Math 1", "Math 2", "Reading 1", "Reading 2"),
          show.ci = FALSE,
          digits = 2,
          show.se=TRUE,
          p.style = "star",
          string.est = "Est.",
          string.se = "SE",
          auto.label = FALSE, 
          pred.labels = c(
            `(Intercept)`="Intercept",
            is_cr="Is CR",
            xsi_c="Difficulty",
            subscaleQuantity="Quantity",
            `subscaleSpace and Shape`="Space and Shape",
            `subscaleUncertainty and data`="Uncertainty and Data",
            `process_shortFormulate` = "Formulate",
            `process_shortInterpret` = "Interpret",
            `process_shortIntegrate and interpret`="Integrate and Interpret",
            `process_shortReflect and evaluate`="Reflect and Evaluate",
            `txt_mix`="Text Mixed/Multiple",
            `txt_non`="Text Non-continuous"
          ),
          rm.terms = c("factor(year)2012","factor(year)2015"))
```

## Table 5

```{r}
dff_ests %>%
  filter(term=="Female X isCR") %>%
  group_by(SUB, YEAR) %>%
  summarise(
    Mean=mean(d_format),
    Median=median(d_format),
    Min=min(d_format),
    Max=max(d_format),
    NSig=sum(B3_p<0.01)
  ) %>%
  pivot_longer(cols = -c("SUB","YEAR")) %>%
  pivot_wider(names_from = c("SUB","YEAR"), values_from="value") %>%
  kable(digits=2) %>% kable_styling()
```

# Figures

## Figure 1

```{r, fig.height=5, fig.width=10}
dff_ests %>%
  filter(term == "Female X isCR") %>%
  mutate(yvar=d_format,
         yvar_se=d_format_se) %>%
  arrange(SUB,YEAR,yvar) %>%
  group_by(SUB,YEAR) %>%
  mutate(
    norder=row_number(),
    yvar_hi=yvar+1.96*yvar_se,
    yvar_lo=yvar-1.96*yvar_se) %>%
  ggplot(aes(x=norder, y=yvar)) +
  geom_hline(yintercept = 0, lty=2) +
  geom_point(aes(pch=factor(CNT=="USA"))) +
  scale_shape_manual(values = c(1,19)) +
  guides(shape="none") +
  geom_errorbar(aes(ymin=yvar_lo, ymax=yvar_hi)) +
  scale_y_continuous(breaks = seq(-0.1,0.4,0.1), limits = c(-0.17, 0.46)) +
  geom_text(aes(y=yvar_hi+0.01, label=CNT, hjust=0), angle=90, size=2) +
  ylab("Standardized Item Format Effect") +
  facet_grid(SUB~YEAR) +
  theme_bw(base_size=14) +
  theme(axis.title.x = element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank())
```

```{r}
ggsave(filename="6-figure1-caterpillarplot.eps", height=5, width=10)
```

How many are negative in math?

```{r}
dff_ests %>%
  filter(term == "Female X isCR") %>%
  group_by(SUB) %>%
  summarise(
    N=n(),
    neg=sum(B3<0),
    negsig=sum(B3<0 & B3_p<0.01),
    possig=sum(B3>0 & B3_p<0.01),
    mneffect=mean(d_format)
  )
```

# Appendix

## Countries and Languages

```{r}
cnt_names <- read.csv(file = "0-CNT-codes.csv")
full_data <- left_join(full_data, cnt_names, by="CNT")
full_data %>%
  group_by(CNT, Country, YEAR, TESTLANG) %>%
  summarise() %>%
  pivot_wider(names_from="YEAR", values_from="TESTLANG") %>%
  kable() %>% kable_styling()
```
